Use the device interface for detecting Triton availability
This allows for each device type to check installed devices for Triton
compatibility and ensure their Triton backend is present.

This PR replaces the has_triton() global method which was previously
used for this task, and moves the initial check for each Inductor
backend on to their associated BaseScheduler subclass. This means that
other backends, such as Halide, can also implement their own
availability checks.

Use the device interface for detecting Triton availability

This allows for each device type to check installed devices for Triton
compatibility and ensure their Triton backend is present.

This PR replaces the has_triton() global method which was previously
used for this task, and moves the initial check for each Inductor
backend on to their associated BaseScheduler subclass. This means that
other backends, such as Halide, can also implement their own
availability checks.

Don't attempt to access non-existent graph in init of MetalScheduling

Restore has_triton as it's used outside of testing