Improve Pareto frontier plot for AutoAC
This was added in https://github.com/pytorch/pytorch/pull/126320. It's a very nice feature, which can be used to predict memory usage for different budget values.

However, it had some limitations, notably in terms of resolution (it only sampled 21 points across the whole range thus missed many threshold values) and in distributed settings.

Here I fix those by using recursive binary searches to identify all thresholds (up to a resolution of 1e-3, which can be made configurable) and output them in SVG (to be able to discern different points), plus I add the rank to the filename and store it in a user-define directory.

ghstack-source-id: 4deba4acffc224ddfc5dc94cc9fd77cde58123ea
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148678