ci: Fix check_binary gcc abi check
All of our binaries should be built with the cxx11-abi now so lets fix
this check to reflect reality.

I also noticed that this particular script is not used widely since this
issue should've been caught in nightlies a long time ago.

Maybe worth an investigation to just remove this script if it's not
actually being used.

Signed-off-by: Eli Uriegas <github@terriblecode.com>

ghstack-source-id: 0f6f94a9877e174d9c025ab5d1e86f7c9f2a6de1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/149104

Signed-off-by: Eli Uriegas <github@terriblecode.com>
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>