Remove some memory overhead in parallel compile workers
Summary: The parallel compile workers are holding on to more memory than they need to because they're loading the compiled modules into memory. Update the post-fork initializer to record when in a subprocess and skip some of the unnecessary overhead.

Test Plan: Ran a test script to compile 15k Triton kernels and used tracemalloc in the subprocs to investigate the overhead. On my devgpu:
* After importing torch in a subproc: 371M
* Without this PR, after compiling 15k kernels: 825M
* With this PR, after compiling 15k kernels: 531M

ghstack-source-id: 4cf29398ea9102c15c03cc467b503ac57eff1ace
Pull Request resolved: https://github.com/pytorch/pytorch/pull/149168