ci: Add sccache to manylinux images
Adds sccache to our manylinux images, these are purposefully built
without the scccache-dist binary since we're not expecting to use that.

Another caveat of these builds is that they are built with the vendored
version of openssl.

This is to set the stage for us to be able to build binaries
seqeuntially.

Signed-off-by: Eli Uriegas <github@terriblecode.com>

ghstack-source-id: ca10e46522973b411b8ae059291ced53a770082e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148419

Signed-off-by: Eli Uriegas <github@terriblecode.com>