[DTensor] Error on illegal view op sharding prop
Following the definition of legal sharding in the previous PR, this
enforces  that view ops raise a clear error message if an invalid
uneven sharding is encountered during view op sharding propagation.

The mechanism is to compute the current local shapes based on the
existing input sharding, and  the new local shapes based on the proposed
output sharding and new global shape.  The new global shape is
pre-existing, but the computation of new local output shape is new and
requires additional infrastructure, since normally we'd just run the
aten operator to get the new shape, and in this case we can't even do
that since we'd risk performing an impossible view (e.g. tensor of shape
(512,) reshaped locally to (384,) in the particular example being
addressed here.

Fixes #143372

ghstack-source-id: 26c48eb8304a30c3551bef13d5ac6be047df6ba4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/149764