[dynamo] Support `torch.Tensor._make_subclass` and tracing through tensor subclass `__new__`
This builds off the previous patch in the stack, and fully fixes
https://github.com/huggingface/diffusers/issues/10795.

Essentially, tensor subclass in the issue uses
`torch.Tensor._make_subclass`, which has a pretty simple shallow-copy
plus type change semantics, as far as Dynamo is concerned. So this patch
adds a polyfill for it.

As a result, this allows us to trace through many user-defined `__new__`
in tensor subclass (it's similar to how we trace through user-defined
`__new__` for `UserDefinedClassVariable`), so this patch also faithfully
trace through these `__new__` methods.

ghstack-source-id: ee4af29c6951f55deea7192d196df71daef63df6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/149483